PB-1252: store file sizes of assets and collection total data size as BigIntegers
Currently there's an issue on PROD when uploading large files. In this case it was 12.5 GB.
The file sizes currently are stored in IntegerFields, which is not enough when storing
file sizes of very large assets in bytes. A DataError integer out of range was raised.
Changing the according fields to BigIntegerFields solved this issue.

+++ app/stac_api/models.py
@@ -309,7 +309,7 @@ class Collection(models.Model):
         "in which the underlying assets data are updated."
     )
 
-    total_data_size = models.IntegerField(default=0, null=True, blank=True)
+    total_data_size = models.BigIntegerField(default=0, null=True, blank=True)
 
     allow_external_assets = models.BooleanField(
         default=False,
@@ -435,7 +435,7 @@ class Item(models.Model):
         "in which the underlying assets data are updated."
     )
 
-    total_data_size = models.IntegerField(default=0, null=True, blank=True)
+    total_data_size = models.BigIntegerField(default=0, null=True, blank=True)
 
     # Custom Manager that preselects the collection
     objects = ItemManager()
@@ -623,7 +623,7 @@ class AssetBase(models.Model):
         "-1 means that the data is not on a regular basis updated."
     )
 
-    file_size = models.IntegerField(default=0, null=True, blank=True)
+    file_size = models.BigIntegerField(default=0, null=True, blank=True)
 
     def __str__(self):
         return self.name
@@ -767,7 +767,7 @@ class BaseAssetUpload(models.Model):
         "This field can only be set via the API."
     )
 
-    file_size = models.IntegerField(default=0, null=True, blank=True)
+    file_size = models.BigIntegerField(default=0, null=True, blank=True)
 
     content_encoding = models.CharField(
         choices=ContentEncoding.choices, blank=True, null=False, max_length=32, default=''
